backend/requirements.txt
fastapi==0.121.1
httpx==0.28.1
pydantic==2.12.4
python_jose==3.5.0
uvicorn==0.38.0
